RW-5820 move links to criteria search component
